Adding the CommunityPost Controller & updating the tables

+++ database/migrations/2020_09_08_173607_create_dislikes_of_reviews_table.php
@@ -14,8 +14,8 @@ class CreateDislikesOfReviewsTable extends Migration
     public function up()
     {
         Schema::create('dislikes_of_reviews', function (Blueprint $table) {
-            $table->integer('user_id');
-            $table->integer('review_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('review_id');
             $table->timestamps();
 
             $table->primary(array('user_id', 'review_id'));
